Work for graphics functionality
Memory is beginning to be a severely limiting factor. The device is
barely stable as it is, and the Serial buffer is being corrupted somehow
when read by both CmdMessenger and Serial.readBytes(). Next steps would
be to create a stripped-down test script that only has the graphics
commands and see if it performs any better.