More work for graphics
Completely overhauled the communication protocol. Dropped CmdMessenger
in favor of something much more lightweight (manual bytes, how fun).
Stripped the Arduino code down to be graphics-only to reduce the memory
cost. This commit includes a working implementation of an 8-bit color
graphics driver that runs reasonably well (10-12 FPS). Next steps are to
add back the original functionality and convert the code to use the
new protocol.