Work for #6 and #8
There are some overlap issues with threads. Need to implement a
queue to avoid crashing the device. There is also an outstanding bug
with the CMD_ACK handler that seems to process a command for every
byte sent by the client, which fills the buffer and messes up other
commands.